Move uppercase filenames to weapons plugin
